Add an option to create a Locale File.
Now it'll be possible to create an locale yaml file from the fixed texts
with the created i18n keys.

To create the file it's just use the  option `--create-locale-file
<file_name>` and a `<file_name>.yml` will be created with all the keys
created, using the "default" texts as values.

For now, the file is created as `en`. Maybe would be nice to add other
option to specify the language.